fix: cleanup and divided data classes for better separation
